requirements.txt
Flask==2.3.3
openai==1.51.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
jsonschema==4.19.0
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.27.0
